+++ src/modules/Si443x/Si443x.cpp
@@ -576,11 +576,11 @@ int16_t Si443x::setDataShaping(uint8_t sh) {
   }
 }
 
-void Si443x::setRfSwitchPins(uint8_t rxEn, uint8_t txEn) {
+void Si443x::setRfSwitchPins(uint32_t rxEn, uint32_t txEn) {
   _mod->setRfSwitchPins(rxEn, txEn);
 }
 
-void Si443x::setRfSwitchTable(const uint8_t (&pins)[Module::RFSWITCH_MAX_PINS], const Module::RfSwitchMode_t table[]) {
+void Si443x::setRfSwitchTable(const uint32_t (&pins)[Module::RFSWITCH_MAX_PINS], const Module::RfSwitchMode_t table[]) {
   _mod->setRfSwitchTable(pins, table);
 }
 
@@ -612,7 +612,7 @@ void Si443x::setDirectAction(void (*func)(void)) {
   setIrqAction(func);
 }
 
-void Si443x::readBit(uint8_t pin) {
+void Si443x::readBit(uint32_t pin) {
   updateDirectBuffer((uint8_t)_mod->hal->digitalRead(pin));
 }
 #endif
